Consolidate file related helper functions
