+++ front-end/src/components/Domain/Domain.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable import/prefer-default-export */
 /*
  * This program has been developed by students from the bachelor Computer Science at
  * Utrecht University within the Software Project course.
@@ -11,55 +12,6 @@
  * @packageDocumentation
  */
 import { Topic } from '@models/Domain';
-import { message } from 'antd';
-import { UploadChangeParam } from 'antd/lib/upload';
-import { UploadFile } from 'antd/lib/upload/interface';
-/**
- * Validate a file to be an .owl file.
- * @param file The file to validate
- */
-const validateOWL = (file: File): boolean => {
-  if (!(file.name.endsWith('.owl') || file.name.endsWith('.xml'))) {
-    message.error(`${file.name} is not an ontology file`);
-    return false;
-  }
-  return true;
-};
-
-/**
- * Handle changes when uploading a file.
- *
- * This function is mostly used to limit the number of files
- * in an Ant Design Upload component to one.
- * @param info Upload changes information
- * @param callback The function to call with the changed file. For example, to update a state.
- */
-const onFileChange = (
-  info: UploadChangeParam<UploadFile<any>>,
-  callback: (updatedList: any[]) => void,
-) => {
-  // Remove wrong uploads: file.status is empty when beforeUpload returns false
-  let updatedList = info.fileList.filter((file) => !!file.status);
-  // Make sure only one file is uploaded
-  if (updatedList.length > 1) {
-    updatedList = [updatedList.pop()];
-  }
-  callback(updatedList);
-};
-
-/**
- * Validate a file to be a .json file.
- * @param file The file to validate
- */
-const validateJSON = (file: File): boolean => {
-  if (file.type !== 'application/json' && !file.name.endsWith('.json')) {
-    message.error(`${file.name} is not a JSON file`);
-    return false;
-  }
-  return true;
-};
-
-export { validateJSON, validateOWL, onFileChange };
 
 /**
  * Fetch all topics from the back-end.
