+++ front-end/src/helpers/Files/ReadFileContent.tsx
