+++ front-end/src/helpers/Files/index.tsx
@@ -1,0 +1,3 @@
+export { default as downloadFile } from './DownloadFile';
+export * from './ReadFileContent';
+export * from './Upload';
